[ALPS07779043] gpu: fix build error
Type error introduced by new clang

MTK-Commit-Id: 929d7a337dd919a32c25c3ba091cd887f4193446

CR-Id: ALPS07779043
Feature: Kernel SI Operation

+++ drivers/gpu/mediatek/gpufreq/v2/gpufreq_debug.c
@@ -1015,7 +1015,7 @@ void gpufreq_debug_init(unsigned int dual_buck, unsigned int gpueb_support,
 	if (shared_status)
 		g_shared_status = shared_status;
 	else {
-		GPUFREQ_LOGE("null gpufreq shared status: 0x%llx", shared_status);
+		GPUFREQ_LOGE("null gpufreq shared status: 0x%llx", (unsigned long long)shared_status);
 		BUG_ON(1);
 	}
